Add a query object for truncating tables.
This adds a bunch of new files I will going to explain now.

First of all the TruncateQuery object is added. This provides only one
method to truncate() a table by itself. All the other needed methods come from
the extended AbstractQuery object. This will add methods to prepare a Prepared Statement,
bind values or parameters to them, execute the value, get the affected rows back and implement
the magic method for casting the object into an string. It will return the query string.

All the methods are defined by an QueryInterface which defines the common methods for all
query objects and a TruncateInterface which only defines the special method of the query -
in this case truncate()

Usage with the old API:
// executes the query
$GLOBALS['TYPO3_DB']->exec_TRUNCATEquery('pages');

// returns the SQL
$GLOBALS['TYPO3_DB']->TRUNCATEquery('pages');

Usage of the new API:
// returns the SQL
$query = $GLOBALS['TYPO3_DB']->createTruncateQuery()
  ->truncate('pages')
  ->getSql();

// executes the query
$query = $GLOBALS['TYPO3_DB']->createTruncateQuery()
  ->truncate('pages')
  ->execute();

$GLOBALS['TYPO3_DB']->TRUNCATEquery() delegates to the new method
$GLOBALS['TYPO3_DB']->truncateQuery(). This uses the new API internally to
create the query.